+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add tests for AWS CodeArtifact support
+
 ## [0.1.0] - 2023-03-05
 
 ### Added